Add documentation on modifying printer firmware
README.md has been repurposed to explicitly lay out both how to build
the libraries here for use on printer firmware and how to get these
files into the firmware on a printer.

Since the most practical way to get modified libraries onto a printer
is via ssh, a script to use the printer API to enable ssh access was
added here.